Refactor: Refactored get function.
The function forst checks for a valid date and returns null if it's not.
It then checks if the entry is a natural date or unix time (unix time still buggy).
If the entry is natural, natural date and unix time are returned.
Next: Return month as a word not a number.