Bug: Fixed bug with converting unix time to natural time.
The manipulation to change unix time to natural time is fixed.
Also a bug with unix time 0 is fixed. Before it returned year 2000.
For some reason 0 was detected as string not a number so the condition in the if statement that checks for numbers is modified to explicitly check for 0.